Cleanup some raman files
