+++ Basestation_Software.Web/Core/Components/RamanGraph.razor
@@ -1,5 +1,5 @@
 ﻿@inject RoveCommService _RoveCommService
-@using static Basestation_Software.Models.RamanGraphType
+@using Basestation_Software.Models.RamanGraph
 
 <div class="card full-height">
    <div class="card-header">
@@ -16,7 +16,6 @@
 
 
 @code {
-
     DataItem[] ramanGraph = new DataItem[2048];
