--- Basestation_Software.Web/Core/Services/RamanGraphService.cs
@@ -1,6 +1,0 @@
-﻿namespace Basestation_Software.Web.Core.Services
-{
-	public class RamanGraphService{}
-	
-
-}
