+++ Basestation_Software.Web/Program.cs
@@ -25,7 +25,6 @@ builder.Services.AddSingleton<GPSWaypointService>();
 builder.Services.AddHttpClient<MapTileService>();
 builder.Services.AddSingleton<MapTileService>();
 builder.Services.AddSingleton<TaskTimerService>();
-builder.Services.AddSingleton<RamanGraphService>();
 
 builder.Services.AddSingleton<RoveCommService>();
 builder.Services.AddHostedService((sp) => sp.GetRequiredService<RoveCommService>());
